pyproject.toml
[tool.poetry]
name = "caption-search"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
youtube_dl = "^2021.6.6"
webvtt-py = "^0.4.6"
requests = "^2.26.0"
Flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
